requirements.txt
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
python-dateutil>=2.8.0
python-dotenv>=0.19.0 